Fix non-halt when sanity checks failed
This commit fixes a bug present in all the runners which causes them to
not actually halt execution when sanity checks fail in some
circumstances.  It was also found that the behaviour across the runners
was not consistent so the relevant code was refactored out into a
protected method in AbstractRunner and the derived runners all changed
to call the same helper method.

This runSanityChecks() method will see if sanity checks are enabled, if
so try to run them and if the necessary sanity checking level is not met
asks the runner to halt