Fix logic error in checkSanity()
A bug was spotted in checkSanity() by Vamsikrishna Annem regarding the
running of sanity checks.  The bug meant that runners weren't actually
checking for failed operations correctly so the sanity checks were
always reported as passing regardless of the actual status of the
operation.

This commit improves the logic to check that the operation was actually
successful and only report successful operations as passing sanity
checks