Allow enforcing absolute URIs
This commits adds an option to control the API behaviour when
queries/updates contain relative URIs.  By default ARQ will resolve this
temporarily against an implicit base but transmit the URIs in relative
form to the actual services that execute the statements.  In some cases
this can lead to unexpected behaviour, a new option is added which
allows relative URIs to be transmitted in absolute form when desired.

A corresponding --ensure-absolute-uris option is added to the CLI