New experimental query summarization mode
When running in summarized mode the normal query operations are replaced
with variants which where applicable wrap the queries in a SELECT
(COUNT(*) AS ?numResult) { } style query.  This has the effect of making
the system being tested do all the work of the query but only requiring
them to send back the count of results thus eliminating network IO costs
from the benchmarking.