Fix a possible IllegalArgumentException that can occur
When using the intelligent mix runner it is possible that all operations
end up getting excluded and none are run.  When this happens the code
would obtain an invalid value for the min and max operation IDs and try
to use this to look up a non-existent operation.  This commit changes
the logic to avoid using these values if they are invalid i.e. if no
operations are run